airflow-docker/requirements.txt
apache-airflow-providers-postgres
psycopg2-binary
huggingface-hub
openlineage-airflow
openlineage-integration-common
openmetadata-ingestion
